Add files via upload
Fixed a bug in the SSF sliders. It should have been using .value not .maximum of the db limit, was causing it to error if you went over the max value.
Added FX3 table level support. This works by getting the table from the command line when Steam was launched.
Cleaned up some code.